+++ docs_src/advanced_usage/adv_usage_003.py
@@ -1,8 +1,6 @@
-@app.post('/login')
+@app.post("/login")
 def login(response: Response):
     ...
-    token = manager.create_access_token(
-        data=dict(sub=user.email)
-    )
+    token = manager.create_access_token(data=dict(sub=user.email))
     manager.set_cookie(response, token)
     return response