+++ docs_src/advanced_usage/adv_usage_005.py
@@ -1,4 +1,2 @@
 # expires after 12 hours
-long_token = manager.create_access_token(
-    data=data, expires=timedelta(hours=12)
-)
+long_token = manager.create_access_token(data=data, expires=timedelta(hours=12))