+++ docs_src/setup/setup_001.py
@@ -1,4 +1,5 @@
 from fastapi import FastAPI
+
 app = FastAPI()
 
 SECRET = "super-secret-key"