+++ docs_src/setup/setup_002.py
@@ -1,4 +1,5 @@
 from fastapi import FastAPI
+
 from fastapi_login import LoginManager
 
 app = FastAPI()