+++ docs_src/usage/usage_002.py
@@ -1,4 +1,4 @@
-@app.post('/login')
+@app.post("/login")
 def login(data: OAuth2PasswordRequestForm = Depends()):
     email = data.username
     password = data.password
@@ -7,10 +7,8 @@ def login(data: OAuth2PasswordRequestForm = Depends()):
     if not user:
         # you can return any response or error of your choice
         raise InvalidCredentialsException
-    elif password != user['password']:
+    elif password != user["password"]:
         raise InvalidCredentialsException
 
-    access_token = manager.create_access_token(
-        data={'sub': email}
-    )
-    return {'access_token': access_token}
+    access_token = manager.create_access_token(data={"sub": email})
+    return {"access_token": access_token}