+++ docs_src/usage/usage_003.py
@@ -1,3 +1,3 @@
-@app.get('/protected')
+@app.get("/protected")
 def protected_route(user=Depends(manager)):
-    return {'user': user}
+    return {"user": user}